Give monsters 4 bonus HP per HD.
I ran some math and simulations.

A BX fighter with +1 STR hits for 1d8+1 damage 50% of the time against leather
(ac 12), so an average damage per round of `5.5*.5 = 2.75`

A 1st level fighter in sovereign with +1 str, +1 stab, and armsmaster attacks
at +3 to hit (so 60% against ac 12) for 1d8+3 damage, with a floor of 5 shock,
so 40% of the time they do 5, and 60% of the the time they do
(5,5,6,7,8,9,10,11) which averages to 7.625, for an expected damage per round
of `.4 * 5 + .6 * 7.625 = 6.575`, which is ~2.4x higher than in BX.

We can also compare parties!

Say that the BX PCs are:

alice: fighter. +1 to hit for 1d8+1 damage
bob: dwarf. +1 to hit for 1d8+1 damage
carol: cleric. +0 to hit for 1d6 damage
david: thief. +0 to hit for 1d6 damage
erin: magic user. +0 to hit for 1d4 damage

It takes this party ~4.33 rounds to defeat a party of 2d4 orcs.

Whereas in sovereign, it looks like:

alice: fighter. +3 to hit for 1d8+3 damage. 5/ac 13 shock
bob: dwarf. +3 to hit for 1d8+3 damage. 5/ac 13 shock
carol: cleric. +2 to hit for 1d6+1 damage. 2/ac 18 shock
david: thief. +2 to hit for 1d8+2 damage
erin: magic user. +3 to hit for 1d6+2 damage

and it only takes ~2.1 rounds to defeat the same party of 2d4 orcs.

Giving each orc an extra 4 hp makes them live ~2.9 rounds which gets this
closer to BX parity.